requirements.txt
fastapi
uvicorn
pandas
scikit-learn
mlflow
shap
google-generativeai
python-dotenv
streamlit
requests